Reset libuv stop_flag after terminal_teardown()
When shuttind down, in event_teardown(), the event loop stop_flag
is set to 0 to prevent the final call to uv_run() to exit early.
However the later call to terminal_teardown() calls tui_stop()
that can also call uv_stop() setting the flag again.

Failling to properly terminate the loop triggers an abort in
event_teardown(), Moved the flag reset after the call to
terminal_teardown() to avoid it - #2466