Ignore virtcols after the 32000th in computations
This is to avoid crashing or hanging when editing a file than contains
ludicrously long lines (more than 100,000,000 virtual columns).

The change is in plines_win_nofold, which is called by wrapping and folding
code. As a result, wrapping and folding may be done incorrectly when the UI is
capable of rendering more than 32000 characters at a time (tiny font).

fixes #2838