tests: Migrate legacy test 73.
The test relies on the default value of 'wildmode' in plain Vim.  So it has to
be set at the top of the test.

The directories and files needed for this test are created with lua functions.
This makes it easier to understand the paths as the old Vim test script used
`:cd` a lot.