tests: Modernize migrated legacy test 73.
The output file is removed from the test.  All checks are done directly by
`expect()`ing the correct files.  Tab characters in the input are escaped as
`\t` in lua strings for readability.  The setting of visualbell is removed as
it did not influence the test.  Lastly the use of separate windows for the
test steps is dropped.