main: Move event_init to before command_line_scan
This fixes a segfault when nvim is started in Arabic mode (-A).

The command_line_scan() function can cause options to be set, including
the 'arabic' option. The 'arabic' option in turn force-sets the keymap
option, which causes the keymap file to be loaded.

However, the arabic keymap file is long enough to cause a breakcheck. At
that point, the event system has not been initialized yet, so the
breakcheck fails with a segfault.

This commit moves the init of the event before command_line_scan().