Replace MAYBE usages with explicit TriState enum
The MAYBE value is used as a kind of 'unknown' value in some
optimizations. This commit makes it more explicit where the trick is
used and makes it clearer that some values look like, but aren't,
boolean.

kTriFalse, kTriTrue and kTriMaybe are equal to 0, 1 and 2 (MAYBE)
respectively, so behaviour should not be affected.