provider: Move Python script host into Python module.
When a :python command is executed in an installed Neovim, script_host.py
is compiled to bytecode, resulting in a script_host.pyc file and a
__pycache__ directory.

A distribution's package manager doesn't know about these compiled files,
though, as this only happens at runtime. The general idea is therefore to
pre-compile Python code to bytecode in the package creation step (i.e. in
or before make install). For the Python module, this is done automatically
by its setup.py. For the main neovim package, doing this is a problem, as
it would require adding a dependency on Python.

Therefore, move the script host into the neovim/python-client project to
allow precompiling.

Closes #3349.