Reproducible builds: Honor compilation date in $SOURCE_DATE_EPOCH
There is an ongoing effort[0] to make FOSS software reproducibly
buildable.  In order to make NeoVim build reproducibly, it is necessary
to allow defining the date/time used in longVersionWithDate as part of
the build process.

This commit enables that by adding support for the SOURCE_DATE_EPOCH
spec[1].  When the $SOURCE_DATE_EPOCH environment variable is defined,
it will be used to define the NVIM_VERSION_BUILD_DATE preprocessor
define.

If NVIM_VERSION_BUILD_DATE is not defined, the existing behavior of
relying on the preprocessor's __DATE__/__TIME__ symbols will be used.

[0]: https://reproducible-builds.org/
[1]: https://reproducible-builds.org/specs/source-date-epoch/