regexp_nfa.c: Speed up find_match_text()
An extra test in commit 0a116c828debc6192a6bfb6bceb8cf020e867db0 was
introduced, to check for end of string with a call to strlen(). This was
necessary, because an incorrect length for invalid byte sequences was
used to step through the string. This slowed down find_match_text()
compared to vim's version.

To speed up things, the extra check was removed and a sequence length
of 1 for invalid byte sequences is used.

Fixes issue #3486