tests: Migrate legacy test 47.
The test was modernized in two ways: First all auxiliary files where written
from lua and second the results are not expected in the test.out file at the
end.  Instead the vimscript variables where the results are redirected to are
expected during the test.  This was done to avoid complications with writing
to the correct buffer after migration.