remove minor version references to python in ci scripts
Usually pip3 and python3 are symlinks to whichever latest minor version
of python 3 are installed. Using specific minor versions like python3.3
breaks when a new minor version is released, and is not necessary unless
the particular version has to be enforced.

Current travis builds are failing because python3.3 can not be found.
The latest release is python3.5.1, and most likely the available version
from brew or repositories changed.