The substitution now works live;
There are still some issues that need to be fixed.
It works well for simple cases.
We implemanted it in such a way that it could be, in theory, applied to any
other function that has a live mode. This live mode would then be called at each
user's input.

We started to implemant some lua tests but they do not work yet.