option: Do not expand options, obtained from XDG vars
It is a wrong thing to do, this makes valid variable values be treated
incorrectly: in

    XDG_DATA_HOME='/home/$foo/.local/share'

`$foo` should be treated literally and not expanded to `foo` environment
variable value.

Also makes option_expand not try to expand too long strings even if these too
long strings are default values. Previously it thought that default values
should always be expanded. Also does not try to expand NULL should it be the
default value just in case.

Fixes #4961