Clean up
Cleaned up several functions that no longer need to be public facing, as
mentioned by justinmk.

Also removed any references to notes, as they really don't fit in with
the general idea of the feature otherwise. Report that something
succeeded or that it did not. One can use info if required.

Removed any references to `diagnose`