runtime/2html.vim: check for termguicolors.
Before this commit, even when using termguicolors, 2html.vim would try
to guess the colours that should be used based on a builtin palette that
was deemed 'good enough', often resulting in hideous results.

This patch makes 2html check for either `gui_running` or `termguicolors`
before deciding to approximate colorus.