Appveyor: Enable Python provider tests
Setup python2/3 and install the Neovim client in Appveyor to
enable the python tests.

We are using the Python instalation provided by Appveyor, because
dependencies pyuv, greenlet have issues compiling with MinGW-w64.
And this way we avoid building those too.

We create a copy of python.exe as python3.exe to workaround the lack
of version specific alias for python, that prevents the python provider
from finding python3.