Add API version information to metadata
The API version number is disconnected from the NVIM version. The
API metadata holds the current API version, and the lowest compatible
version supported by this instance.

Release 0.1.6 will be the first release holding API version info, as follows

    metadata['api_level'] = {current: 1, compatibility: 0}

In unreleased versions of the API an additional attribute is available

    metadata['api_level'] = {current: 1, compatibility: 0, prerelease: true}

The version may remain unchanged across Neovim releases if the
API has not changed.

When changing the API the CMake variable NVIM_API_PRERELEASE is set to
true, and  NVIM_API_CURRENT/NVIM_API_COMPATIBILITY are incremented
accordingly.